requirements.txt
anyio==4.1.0
APScheduler==3.6.3
black==23.11.0
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
emoji==2.9.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==0.20.0
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tornado==6.4
tzlocal==5.2
urllib3==2.1.0
